Further PR feedback
