+++ hack/generator/pkg/codegen/code_generator.go
@@ -128,9 +128,8 @@ func createAllPipelineStages(idFactory astmodel.IdentifierFactory, configuration
 
 		assertTypesCollectionValid(),
 
+		// TODO: We need better testing for the below
 		removeEmbeddedResources().UsedFor(ARMTarget), // TODO: For now only used for ARM,
-
-		removeEmptyObjectTypes(),
 
 		// Apply export filters before generating
 		// ARM types for resources etc:
