+++ hack/generator/pkg/codegen/golden_files_test.go
@@ -154,7 +154,7 @@ func NewTestCodeGenerator(testName string, path string, t *testing.T, testConfig
 			// removeEmbeddedResources treats the collection of types as a graph of types rooted by a resource type.
 			// In the degenerate case where there are no resources it behaves the same as stripUnreferenced - removing
 			// all types. Remove it in phases that have no resources to avoid this.
-			codegen.RemoveStages("removeEmbeddedResources", "removeEmptyObjectTypes")
+			codegen.RemoveStages("removeEmbeddedResources")
 			codegen.ReplaceStage("stripUnreferenced", stripUnusedTypesPipelineStage())
 		}
 	case config.GenerationPipelineCrossplane:
